+++ SortingAlgorithm.java
@@ -6,5 +6,5 @@
  */
 public interface SortingAlgorithm {
 
-	void sort(double [] a);
+	void sort(int [] a);
 }
